optionally allow using mtools to build the hdd file
This avoids the need for sudo with the current approach, but requires mtools to be installed

+++ GNUmakefile
@@ -71,9 +71,15 @@ $(IMAGE_NAME).hdd: limine kernel
 	rm -f $(IMAGE_NAME).hdd
 	dd if=/dev/zero bs=1M count=0 seek=64 of=$(IMAGE_NAME).hdd
 	parted -s $(IMAGE_NAME).hdd mklabel gpt
-	parted -s $(IMAGE_NAME).hdd mkpart ESP fat32 2048s 100%
+	parted -s $(IMAGE_NAME).hdd mkpart ESP fat32 1MiB 100%
 	parted -s $(IMAGE_NAME).hdd set 1 esp on
 	./limine/limine bios-install $(IMAGE_NAME).hdd
+ifeq ($(USE_MTOOLS),yes)
+	mformat -i $(IMAGE_NAME).hdd@@1M
+	mmd -i $(IMAGE_NAME).hdd@@1M ::/EFI ::/EFI/BOOT
+	mcopy -i $(IMAGE_NAME).hdd@@1M kernel/kernel.elf limine.cfg limine/limine-bios.sys ::/
+	mcopy -i $(IMAGE_NAME).hdd@@1M limine/BOOT*.EFI ::/EFI/BOOT
+else
 	sudo losetup -Pf --show $(IMAGE_NAME).hdd >loopback_dev
 	sudo mkfs.fat -F 32 `cat loopback_dev`p1
 	mkdir -p img_mount
@@ -86,6 +92,7 @@ $(IMAGE_NAME).hdd: limine kernel
 	sudo umount img_mount
 	sudo losetup -d `cat loopback_dev`
 	rm -rf loopback_dev img_mount
+endif
 
 .PHONY: clean
 clean:
